# optimaze concurrent read performance