--- a/file_manager.go
+++ b/file_manager.go
@@ -41,12 +41,7 @@ func (fm *fileManager) getDataFile(path string, capacity int64) (datafile *DataF
 		}
 	}
 
-	return &DataFile{
-		path:       path,
-		writeOff:   0,
-		ActualSize: 0,
-		rwManager:  rwManager,
-	}, nil
+	return NewDataFile(path, rwManager), nil
 }
 
 // getFileRWManager will return a FileIORWManager Object