For now, we have four steps for reading a entry from disk: read meta,  read bucket, read key, and read value. I think when we had read meta, we already konw the total length of bucket, key, and value, and the location of bucket, key, value is continuous in disk.So we can read bucket, key, value once, it can reduce syscall times from 4 to 2.
And the memery of meta can reuse, bucause it will parse into a object in go struct after it read from disk, so the buffer it malloc can be reuse in next time to read meta from disk.
Besides, I also write a benchmark test about this code change.
I also build a database which has 100 keys and different size value in local, and concurreny read it in 10 goroutines. 
And the result as below, first item is master performance, and the other is this branch

when value size is 50:
`BenchmarkConcurrentRead-10    	    2054	    663687 ns/op	   69053 B/op	    1303 allocs/op`
`BenchmarkConcurrentRead-10    	    3390	    406251 ns/op	   66724 B/op	    1103 allocs/op`

when value size is 256:
`BenchmarkConcurrentRead-10    	    1581	    745181 ns/op	   88255 B/op	    1303 allocs/op`
`BenchmarkConcurrentRead-10    	    4075	    395201 ns/op	   87511 B/op	    1103 allocs/op`

when value size is 512:
`BenchmarkConcurrentRead-10    	    1627	    759430 ns/op	  113887 B/op	    1303 allocs/op`
`BenchmarkConcurrentRead-10    	    2794	    404143 ns/op	  116339 B/op	    1103 allocs/op`

As we can see, this is a significant optimization.